57-네비게이터와 메뉴바 수정
