+++ client/teamplay/package.json
@@ -32,7 +32,8 @@
     "react-native-screens": "^3.25.0",
     "react-native-shadow-2": "^7.0.8",
     "react-native-svg": "^13.13.0",
-    "react-native-vector-icons": "^10.0.0"
+    "react-native-vector-icons": "^10.0.0",
+    "react-navigation": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
